+++ cmd/containerd/builtins_linux.go
@@ -23,5 +23,5 @@ import (
 	_ "github.com/containerd/containerd/runtime/v2"
 	_ "github.com/containerd/containerd/runtime/v2/runc/options"
 	_ "github.com/containerd/containerd/snapshots/native"
-	_ "github.com/containerd/containerd/snapshots/overlay"
+	_ "github.com/containerd/containerd/snapshots/overlay/plugin"
 )
