+++ snapshots/overlay/plugin/plugin.go
@@ -18,8 +18,40 @@
 
 package overlay
 
+import (
+	"errors"
+
+	"github.com/containerd/containerd/platforms"
+	"github.com/containerd/containerd/plugin"
+	"github.com/containerd/containerd/snapshots/overlay"
+)
+
 // Config represents configuration for the overlay plugin.
 type Config struct {
 	// Root directory for the plugin
 	RootPath string `toml:"root_path"`
+}
+
+func init() {
+	plugin.Register(&plugin.Registration{
+		Type:   plugin.SnapshotPlugin,
+		ID:     "overlayfs",
+		Config: &Config{},
+		InitFn: func(ic *plugin.InitContext) (interface{}, error) {
+			ic.Meta.Platforms = append(ic.Meta.Platforms, platforms.DefaultSpec())
+
+			config, ok := ic.Config.(*Config)
+			if !ok {
+				return nil, errors.New("invalid overlay configuration")
+			}
+
+			root := ic.Root
+			if config.RootPath != "" {
+				root = config.RootPath
+			}
+
+			ic.Meta.Exports["root"] = root
+			return overlay.NewSnapshotter(root, overlay.AsynchronousRemove)
+		},
+	})
 }
